feat:进一步完善相关功能

+++ eva-infra/src/main/java/edu/cuit/infra/gateway/impl/eva/EvaDeleteGatewayImpl.java
@@ -3,6 +3,7 @@ package edu.cuit.infra.gateway.impl.eva;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.update.UpdateWrapper;
 import edu.cuit.domain.gateway.eva.EvaDeleteGateway;
+import edu.cuit.infra.dal.database.dataobject.course.CourseDO;
 import edu.cuit.infra.dal.database.dataobject.eva.CourOneEvaTemplateDO;
 import edu.cuit.infra.dal.database.dataobject.eva.EvaTaskDO;
 import edu.cuit.infra.dal.database.dataobject.eva.FormRecordDO;
@@ -11,9 +12,12 @@ import edu.cuit.infra.dal.database.mapper.eva.CourOneEvaTemplateMapper;
 import edu.cuit.infra.dal.database.mapper.eva.EvaTaskMapper;
 import edu.cuit.infra.dal.database.mapper.eva.FormRecordMapper;
 import edu.cuit.infra.dal.database.mapper.eva.FormTemplateMapper;
+import edu.cuit.zhuyimeng.framework.common.exception.QueryException;
+import edu.cuit.zhuyimeng.framework.common.exception.UpdateException;
 import lombok.RequiredArgsConstructor;
 import org.apache.ibatis.annotations.Update;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -27,39 +31,40 @@ public class EvaDeleteGatewayImpl implements EvaDeleteGateway {
     private final FormTemplateMapper formTemplateMapper;
     private final CourOneEvaTemplateMapper courOneEvaTemplateMapper;
     @Override
+    @Transactional
     public Void deleteEvaRecord(List<Integer> ids) {
         for (Integer id : ids) {
             UpdateWrapper<FormRecordDO> formRecordWrapper = new UpdateWrapper<>();
             formRecordWrapper.eq("id", id);
-            formRecordMapper.delete(formRecordWrapper);
-            //获取评教记录对应的id
-            FormRecordDO formRecordDO=formRecordMapper.selectById(id);
-            Integer taskId=formRecordDO.getTaskId();
-            //删除对应的任务
-            UpdateWrapper<EvaTaskDO> evaTaskWrapper =new UpdateWrapper<>();
-            formRecordWrapper.eq("task_id",taskId);
+            if(formRecordWrapper==null){
+                throw new QueryException("可怜的人类，并未找到找到相应课程");
+            }else {
+                formRecordMapper.delete(formRecordWrapper);
+            }
         }
         return null;
     }
-
+    //ok
     @Override
+    @Transactional
     public Void deleteEvaTemplate(List<Integer> ids) {
-        //TODO 怎么知道评教表单有没有用过
-
         for(Integer id : ids){
-            //模板和快照模板id一样？(假如一样)
-            QueryWrapper<CourOneEvaTemplateDO> courOneEvaTemplateWrapper =new QueryWrapper<>();
-            courOneEvaTemplateWrapper.eq("id",id);
+            //没有分配在课程中
+            QueryWrapper<CourseDO> courWrapper =new QueryWrapper<>();
+            courWrapper.eq("templateId",id);
             //获取对应课程id
-            CourOneEvaTemplateDO courOneEvaTemplateDO=courOneEvaTemplateMapper.selectById(id);
-            Integer courseId=courOneEvaTemplateDO.getCourseId();
-            //判断是不是在课程里面有
-            if(courseId==null){
-                UpdateWrapper<FormTemplateDO> formTemplateWrapper =new UpdateWrapper<>();
-                formTemplateWrapper.eq("id",id);
-                //删除模板
-                formTemplateMapper.delete(formTemplateWrapper);
-            }//TODO 没有的记得报错
+            if(courWrapper==null){
+                UpdateWrapper<FormTemplateDO> formTemplateWrapper = new UpdateWrapper<>();
+                formTemplateWrapper.eq("id", id);
+                if(formTemplateWrapper==null){
+                    throw new QueryException("可怜的人类，并未找到找到相应课程");
+                }else{
+                    //删除模板
+                    formTemplateMapper.delete(formTemplateWrapper);
+                }
+            }else{
+                throw new UpdateException("可怜的人类，该模板已经被课程分配，无法再进行删除");
+            }
         }
         return null;
     }
